Add logic to handle not deploying when an environment does not exist.

+++ providers/pantheon/pantheon-dependencies.sh
@@ -6,19 +6,33 @@
 #
 # "Declarations" of the variables this script assumes
 #
-#declare=${VARNAME:=}
+declare=${PANTHEON_TOKEN:=}
+
+if ! [ -f ~/cache/terminus/bin/terminus ]; then
+    #
+    # Installing Terminus for Pantheon
+    #
+    announce "...Installing Terminus for Pantheon"
+    cd ~/cache
+    rm -f terminus.tar.gz
+    rm -rf terminus
+    wget -O terminus.tar.gz https://github.com/pantheon-systems/terminus/archive/1.6.0.tar.gz
+    tar -xzf terminus.tar.gz
+    rm terminus.tar.gz
+    mv terminus-1.6.0 terminus
+    cd terminus
+    composer update
+fi
 
 #
-# Installing Terminus for Pantheon
+# Symlinking Terminus for Pantheon
 #
-announce "Installing Terminus for Pantheon"
-cd /tmp
-rm -f terminus.tar.gz
-rm -rf terminus
-wget -O terminus.tar.gz https://github.com/pantheon-systems/terminus/archive/1.6.0.tar.gz
-tar -xzf terminus.tar.gz
-mv terminus-1.6.0 terminus
-exit
+announce "...Symlinking Terminus for Pantheon"
+sudo ln -s ~/cache/terminus/bin/terminus /usr/local/bin/terminus
 
-
+#
+# Authenticating Terminus for Pantheon
+#
+announce "...Authenticating Terminus for Pantheon"
+terminus auth:login --machine-token="${PANTHEON_TOKEN}"
